+++ providers/capa/private/capa_test.go
@@ -1,0 +1,15 @@
+package standard
+
+import (
+	. "github.com/onsi/ginkgo/v2"
+
+	"github.com/giantswarm/cluster-test-suites/internal/common"
+)
+
+var _ = Describe("Common tests", func() {
+	common.Run(&common.TestConfig{
+		AutoScalingSupported: false,
+		BastionSupported:     false,
+		ExternalDnsSupported: false,
+	})
+})