+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- CAPA private cluster tests
+- 
 ### Changed
 
 - Remove CPI config from CAPVCD tests so that we use the chart defaults instead.